# user_service/user_service_swagger.yml
swagger: "2.0"
info:
  title: User Service API
  version: "1.0.0"
  description: API for managing users and their friendships

host: localhost:5001
basePath: /
schemes:
  - http

paths:
  /users/:
    post:
      tags:
        - Friends
      summary: Create a new user
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - username
              - password
            properties:
              username:
                type: string
                example: johndoe
              password:
                type: string
                example: mysecurepassword
      responses:
        201:
          description: User successfully created
          schema:
            type: object
            properties:
              message:
                type: string
                example: User created
              user_id:
                type: integer
                example: 1
        400:
          description: Missing username or password
          schema:
            type: object
            properties:
              error:
                type: string
                example: username and password are required
        409:
          description: Username already exists
          schema:
            type: object
            properties:
              error:
                type: string
                example: Username already exists

  /users/{user_id}:
    get:
      tags:
        - Friends
      summary: Get a user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
          description: ID of the user to retrieve
      responses:
        200:
          description: User information
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 1
              username:
                type: string
                example: johndoe
        404:
          description: User not found
          schema:
            type: object
            properties:
              error:
                type: string
                example: User not found

  /users/{user_id}/friends/:
    post:
      tags:
        - Friends
      summary: Add a friend to the user
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
          description: ID of the user adding a friend
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - friend_id
            properties:
              friend_id:
                type: integer
                example: 2
      responses:
        201:
          description: Friend successfully added
          schema:
            type: object
            properties:
              message:
                type: string
                example: Friend added
        200:
          description: Friends are already friends
          schema:
            type: object
            properties:
              message:
                type: string
                example: Already friends
        400:
          description: Invalid request
          schema:
            type: object
            properties:
              error:
                type: string
                example: Cannot add yourself as a friend
        404:
          description: One or both users not found
          schema:
            type: object
            properties:
              error:
                type: string
                example: User or friend not found

    get:
      tags:
        - Friends
      summary: Get a user's friends
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
          description: ID of the user whose friends to retrieve
      responses:
        200:
          description: List of user's friends
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                  example: 2
                username:
                  type: string
                  example: janedoe
        404:
          description: User not found
          schema:
            type: object
            properties:
              error:
                type: string
                example: User not found

  /login:
    post:
      tags:
        - Auth
      summary: Log in with username and password
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - username
              - password
            properties:
              username:
                type: string
                example: johndoe
              password:
                type: string
                example: mysecurepassword
      responses:
        200:
          description: Login successful
          schema:
            type: object
            properties:
              message:
                type: string
                example: Login successful
              user_id:
                type: integer
                example: 1
        400:
          description: Missing username or password
          schema:
            type: object
            properties:
              error:
                type: string
                example: username and password are required
        401:
          description: Invalid credentials
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid credentials